requirements.txt
pyTelegramBotAPI>=4.14.0
openai>=1.0.0
# Real Haystack integration for the "History Day" feature
haystack-ai>=2.26.1
# Инструмент агента через LangChain @tool + convert_to_openai_tool
langchain-core>=0.3.0
python-dotenv>=1.0.0
httpx>=0.23.0
requests>=2.28.0
yandex-cloud-ml-sdk>=0.17.0
logtail-python>=0.3.0
# Optional: for knowledge base document parsing
docling>=2.82.0
pypdf>=4.0.0
python-docx>=1.1.0
# Report generation (PDF from conversation history)
jinja2>=3.1.0
weasyprint>=60.0
# Vector database for quotes collection
lancedb>=0.20.0
pyarrow>=14.0.0